adjust time_stamp for file: sample_41_Scene_Understanding_26-50.mp4
The time_stamp annotation is incorrect. The file sample-41_Cene_Understanding-26-50.mp4 has a duration of 5:17, and the original time_stamp is "00:06:22", which exceeds the file duration and causes the eval.py code to report an error. According to the content of the question, adjust the value of time_damp to "00:04:22"